ci-operator: pre-censor fields before encoding output
When user secrets end up in our output files *and* they include
characters that are invalid in the encoding we're going to use to write
those files (like JSON or XML), we end up not censoring anything since
the encoded content we run the censorer over does not contain the actual
secret data. Pre-censoring fields in these structs is fragile, but a
reasonable compromise to writing custom, complex and also fragile
reflection-based implementations that would attempt to operate over
different structs generically. When we use `api.SaveArtifact` we must
remember to pre-censor as necessary beforehand.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>