DPTP-2262: Support importing cluster claim release streams
This PR adds support for importing the release streams of cluster claims
for tests that use cluster claims. To avoid conflicts that can result in
naming collisions, the test step with the cluster claim creates an
import step using the name provided by the user appended by the name of
the test.

The `DependencyParts` function has been modified to accept a struct
called `ClaimRelease` containing this name (`ReleaseName`) and the name
of the release that it overrides (`OverrideName`). If the release stream
for the `OverrideName` in the `ClaimRelease` matches the imagestream
that the user requests, `DependencyParts` returns the image stream for
the release in `ClaimRelease.ReleaseName`. `DependencyParts` is used
extensively across the project when resolving image streams, which
simplifies overriding image names across the project.