pod-scaler: add e2e tests that use a real Prometheus
In these tests (and the local running mode for the pod-scaler in
general), we start by filling a Prometheus server with some fake data
using the new backfill command from `promtool`. Then, we run the
pod-scaler producer and assert something about the data we've fetched.
We cannot sensibly assert about the rates, so for now we just assert
about memory usage, but this should suffice as the content of a query
does not change how we get the data from Prometheus and store it. In the
time it took to write this e2e test, the following issues were resolved:

 - the promtool utility is extremely slow, we use the latest possible
   version now to take advantages of new patches I wrote to increase
   the speed it runs by 100x
 - there were two off-by-one errors in the way that we queried
   Prometheus which would lead to double-counted data points in rare
   cases

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>