pod-scaler: improve boundary conditions on query range calls
There's no way to query the data in Prometheus at the native resolution,
which means that we need to be really careful to not send queries which
have edges that may overlap or which are so short as to not obviously
contain novel data. Prometheus will do an instant query evaulation at
the start time of the range as well as each multiple of the step until
the evaulation time exceeds the end time of the range, and each instant
query evaulation will pick up the most recent value. Therefore, if a
query start time is closer than a step length to the end of a previous
query, it will just pick up the previous end value. Similarly, if we
query for a time period that's not longer than one step we are not
guaranteed to get any novel data.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>