CI Operator: Make sure cloning is not disabled when cloning
Currently, the decoration config the ci-operator gets gets passsed on to
pods it creates. In many cases, it has skip_cloning: true set which
means the decoration utils code will create an empty clonerefs
container which currently leads to a NPD.

We only attempt to create a clonerefs container if we definitely want
cloning, so we need to overwrite that setting.