ci-operator: don't use the root cache when building new caches
We label the output images we build with the versions of the inputs that
were used to build them, in order to be able to detect when an input has
changed so that we do not use a cached image in error. When we build on
top of a cache, however, we end up recording the cache as the input,
since we're built on top of it. However, we only know how to check for
input changes from the actual root. By not building on top of caches
when we're about to promote new caches, we side-step this issue.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>