ci-operator: wait for 30min before giving up on Pods
Ultimately, our hard time-out on waiting for a Pod to reach the Running
phase was written to catch issues like ImagePullBackOff which do not
terminate the Pod but do cause it to hang forever waiting for e.g.
credentials. This is a defense mechanism to ensure that jobs finish
running and not time out themselves. We do not need it to be very
efficient. We are currently killing Pods that took a while to start the
Pod sandbox with an overly agressive timeout.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>