Add method to inject a test from one ci-op config to another
`WithPresubmitFrom()` returns a new configuration, where a selected test
from the source configuration is injected into the base configuration,
together with all elements from the source configuration that are
potentially necessary to allow that test to function in the context of
the base configuration. The intended use case is to inject the test
definition of a "release job" (informing/blocking) into a component
ci-operator config to allow dynamically executing any such job on any
component PR, without the need to clutter each individual component
ci-operator config.

This commit is a part of an experimental effort and the code is not used
anywhere in the production path, so it intentionally does not deal with
all corner cases stemming from different ci-op config shapes on both
sides, conflicts etc.