pkg/validation: validate targets are unique
One unfortunate aspect of the current static validation is that it is
done on the unprocessed input configuration.  This results in
significant duplication between validation and parsing code.  This
change attempts to move to a more tenable position by performing
validation on the graph configuration structures, which are much closer
to the final execution graph.

https://issues.redhat.com/browse/DPTP-2388 is implemented in this change
as well as a PoC.

---

Use the naming methods and graph configuration data added in the
previous commits to validate that all targets in a `ci-operator`
configuration are unique.

Add this validation to the configuration loading process in
`ci-operator` and make `checkconfig` build and validate the static
portion of the graph.