Vault subpath proxy: Allow adding new values to secrets
Currently, the proxy would refuse all changes where a key in a vault
entry targets a kube secret that is already targeted. This unfortuantely
included the case where the key from the same vault entry got another
key, making it impossible to add additional keys to vault entries.

This change fixes that.

Ref https://issues.redhat.com/browse/DPTP-2416