openstack: split "mecha" into two profiles & slices
Our OpenStack cloud used for CI, named `mecha`, has now two availability
zones so we would like to have two different cluster profiles and quota
slices.

For that, we are adding them as new profiles, but we'll remove the
deprecated `openstack-vh-mecha` profile once we have switched our jobs
to use the new ones.