pkg: Allow alternative reference-step command file extensions
2a31056baf (multi_stage: write commands.sh to configmap, 2021-03-10, #1776)
pivoted to directly executing command files allowing non-shell via
shebangs.  But writing Python in a file with a *.sh extension isn't
all that intuitive, and also confuses some editors.  This commit
softens things to allow any extension.

Ideally we'd load the reference YAML and pull commands out of there
[1].  This would require some rewrites, since a bunch of these
existing handlers are file-based, with GetChangedRegistrySteps logic
like:

  1. Get a set of changed files.
  2. Iterate over the files to see if they seem to match a step.
  3. Return any identified steps.

Instead we'd want:

  1. Get a set of changed files.
  2. Iterate over the nodes in the registry.
  3. For each node, get the set of files consumed (e.g. both the ref
     YAML and command file for a reference step).
  4. Mark the node as changed if there is any overlap between the
     changed-files set and the node-consumed-files sets.
  5. Return any identified steps.

But that's a bigger pivot, so punting for now.

[1]: https://github.com/openshift/release/blob/66ee37c6805b900b1dc1d537126956b352286cbb/ci-operator/step-registry/ipi/install/install/ipi-install-install-ref.yaml#L5