Makefile: update diffing targets to honor dummy file
We want to keep running the linters on raw source code, but they fail
when the files that will be embeded via `//go:embed` do not exist.
Therefore, we check in dummy files that we must remove when building our
real servers. Since we use the output of the build as our cache, cached
runs will not have the dummy files and the unbounded `git diff` will
fail in the `verify-gen` target. This is getting complicated/Rube-
Goldberg-eque, but if we need lint to run before build ...

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>