sprint-automation: search for more issues to intake
We previously only looked at issues filed in the last week, as we had
never done intake before and did not want to swamp people with long
lists. We've had intake for a while now and do not want anything to fall
in the cracks, so add 30d here to effectively give three intake shifts a
chance to move a card to ready.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>