Refactor Prowgen code for `PodSpec` for reusability
This change extracts all code that generates `PodSpec` instances that
run `ci-operator` into a Builder pattern generator. The generator
contains individual separate methods that enable a certain trait of the
executed CI job ("uses claims", "runs a template test", etc.) in a way
that is always self-consistent (for example: add a volume, mount it, and
pass a matching option to ci-operator).

The generator encapsulates what content is generated in the PodSpec, but
does no decisions about *when* are different traits used: that is left
on the `GenerateJobs` level.

The change looks scary, but most of the content are test fixtures.