ci-op validation: improve tests and fix nits
In a previous commit I added a new validation, added a test for it, removed the check to see the red test, but found the test actually passes because the testcase is invalid for a different reason. So I imporve the tests to search for an actual one error it wants to see for a testcase, tolerating *other* tests.