pod-scaler: fix a round-tripping bug in the UI
We only want to show a minimal set of fields to the UI for selecting the
correct background data, allowing the views to omit things that are not
important in the context of a specific type of data (pods, RPMs, etc).
We need to ensure that all of these transforms round-trip correctly.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>