promotion: correctly handle multiple promotion names
Promotion correctly tracks multiple promotion destinations for a name, but previously we compressed the items into a map keyed by source, so each source only had a single destination in promotion mapping, leading to some destinations to be silently skipped.

Key the interface map with destinations, so that we can have multiple destinations with the same source.