prowgen: (further) refactor code for reusability
The motivation for this refactor is allowing external code
(payload-testing controller, specifically) to reuse Prowgen code without
needing to duplicate parts of it.

The refactor is based on extracting the bits that handle generating the
`JobBase` part of the ProwJob, and making two constructors for new
builder: one that covers the part of the `JobBase` that is generated
from ci-operator config as a whole, and one that covers the parts that
are generated for individual test items in that config. This allows to
simplify the signatures of the `Generate*ForTest` methods, because they
now only handle the Presubmit/Postsubmit/Periodic generation, without
the `JobBase` part they have in common.

The refactor could go even further, there are still duplicities to
eliminate, but I would like to stop here, because this PR carries the
changes necessary for integrating
https://github.com/openshift/ci-tools/pull/2462 work into the
controller. I believe that even in this partial way it improves the
structure of Prowgen code by simplifying some of the huge methods that
were there.