pkg/registry: ignore errors for overridden chains
This change makes validation accept a tricky corner case with workflows:

- a step declares a parameter (e.g. the `openshift-e2e-test` step
  declares `TEST_SKIP`)
- a workflow sets a value for that parameter (e.g. the
  `openshift-e2e-vsphere` workflow uses `TEST_SKIP` to always skip a
  suite in vSphere tests)
- a test uses the workflow, but replaces the chain that contains the
  step (e.g. the `e2e-vsphere-operator` uses `openshift-e2e-vsphere`
  with its own `test` chain)

Currently, this results in a validation error since the resulting
configuration's `env` section, inherited from the workflow, contains a
parameter that is not declared in any step.  It cannot be cleared in the
test's `env` section since mapping fields are merged, not overridden,
during resolution.  The only solution is to superfluously declare the
parameter in one of the tests' steps.

This change makes the resolution process keep track of the overridden
steps so that this type of error can be ignored.  This works since:

1. this situation can only occur in test/workflow combinations: in no
   other place where parameters are declared/overridden can a chain be
   replaced
2. parameter resolution is done bottom-up: ignored parameters are not
   included in the final configuration anyway

Ignoring errors instead of preemptively removing the parameter
declarations that target the overridden steps has the nice benefit that
only the error path is taxed.