prpqr reconciler: label ProwJobs with hashed names
Names of the jobs are unsuitable label values because of their length.
PRPQR reconciler used labels with job names as a way to identify jobs
spawned for the given PRPQR, but because job names can be longer, this
was not reliable (if we just trim, we can get conflicts).

Hence, use the MD5 hash of the name as the value of the label. That way
it can be reliably used for searching without conflicts. Keep the full
job name in an annotation, instead.