prpqr: base tests on data fixtures instead of code
Most of our newer tests that work with more complex structures use data
fixtures instead of code: they are easier to maintain (running with
UPDATE=true updates them when needed) and they are also really good in
inspecting changes.

The only change in the actual tests is that *fewer* members are ignored.
Previously, the `Spec` was ignored when comparing Prowjobs for no
apparent reason, and not ignoring it will improve the tests for the
future when we change how the submitted Prowjobs exactly look like.