pkg/util: add `openat(2)`-based walk function
Replace `filepath.WalkDir` in the configuration and step registry
loading functions with a reimplementation that always opens files
relative to their parent directory.

This should make sure the traversal is consistent even if the root
directory is moved, e.g. when it is a Kubernetes mount that is updated
while the directory tree is being traversed.

The implementation strives to stay as close to the stdlib version as
possible:

https://go.googlesource.com/go.git/+/refs/tags/go1.17.6/src/io/fs/walk.go