ClusterPools: Log claim & cluster name
Include the namespaced name of the ClusterClaim and the name of the
claimed cluster in the info log of a job. Example before:

```
INFO[2022-01-25T17:22:44Z] Waiting for the claimed cluster to be ready.
INFO[2022-01-25T18:11:30Z] The claimed cluster is ready after 48m46s.
```

Example after:

```
INFO[2022-01-25T17:22:44Z] Waiting for the cluster claim the-namespace/the-claim to be fulfilled.
INFO[2022-01-25T18:11:30Z] The claimed cluster my-pool-abc123 is ready after 48m46s.
```