Further simplify filtering PRs with failed required jobs
I realized we can simplify the code even further: we do not need to
filter the jobs to the ones that *should* trigger, we can easily broaden
the search to the jobs that *could* trigger, which is a much simpler
check (it does not need the inspection of the actual changed files).

If a required job was triggered even manually, then it needs to pass:
the retester does not need to care. If it sees a context that belongs to
a required job, then the job can (and should) be apparently retested.