ci-operator: do not hide reporting log line
Partially reverts b00a10669f79c36c8f6bd4d4e87eb3a905c6ede7.

While not generally of interest to the regular user, hiding this line
makes it impossible to find instances of a particular error reason using
https://search.ci.openshift.org.

The original messages were in the following format:

```
Reporting job state 'succeeded'
```

```
Reporting job state 'failed' with reason 'the:reason'
```

which is not terrible for a user-facing message, so I'm just reverting
the level back to `Info`.

In the future, we may finally work on presenting results in a better
format.  In the meantime, simply removing the message achieves very
little.