ci-operator: always set security context
d31d564f0 started adding security contexts to multi-stage tests which
use VPN connections, but there is never a reason to execute containers
as UID 0 outside of that case.

Set a security context for all pods created during the execution of a
test.  This includes those that import releases etc. but not builds —
which are `privileged` anyway.

The settings in the `PodSecurityContext` propagate and are overridden by
those in the container's `SecurityContext`, so small changes are
required to the existing ones:

https://kubernetes.io/docs/tasks/configure-pod-container/security-context/#set-the-security-context-for-a-container

This will also make me less nervous when making changes to the VPN
implementation.