ci-operator-configresolver: log less in client
When no logger is passed to the retryable HTTP library, it prints
directly to stdout whenever a request is made. Since we've taken such
great pains to be careful about what output `ci-operator` actually
prints to the log, we should pass in a `logrus` adapter to the library
and print only when necessary.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>