pkg/steps: add RPM repositories for all `refs`
The `RPM_REPO_*` variables are currently exposed based on either the
`refs` _or_ `extra_refs` field of the Prow _job spec_.  This works for
(all examples abbreviated):

- Pre-/post-submit jobs, where the repository is in `refs`, e.g.:
  https://prow.ci.openshift.org/view/gs/origin-ci-test/pr-logs/pull/27320/pull-ci-openshift-origin-release-3.11-e2e-gcp/1551900345024647168

    "spec": {"refs": {"org": "openshift", "repo": "origin"}}

- Periodic jobs, where the repository is in `extra_refs`, e.g.:
  https://prow.ci.openshift.org/view/gs/origin-ci-test/logs/periodic-ci-openshift-origin-release-3.11-e2e-gcp/1552572811980050432

    "spec": {"extra_refs": [{"org": "openshift", "repo": "origin"}]}

However, it does not work for rehearsals, where `refs` is
`openshift/release` and the target is in `extra_refs`, e.g.:
https://prow.ci.openshift.org/view/gs/origin-ci-test/pr-logs/pull/openshift_release/26155/rehearse-26155-pull-ci-openshift-origin-release-3.11-e2e-aws/1552262021293019136

    "spec": {
        "refs": {"org": "openshift", "repo": "release"},
        "extra_refs": [{"org": "openshift", "repo": "origin"}]
    }

---

With this change, `refs` and all `extra_refs` are added as aliases to
the same repository, so that all cases are satisfied.  Extra entries in
the parameter map should be harmless, as they are only used to match
template parameters.  Unused entries should have no effect.