Allow setting always_run in the ci-operator config
If there is an existing job definition then the prow job generation
preserves the value of the always_run setting. However, when generating
a new job it is always set to true unless forced to false by a
run_if_changed or skip_if_always_changed directive in the config.

The result of this is that when copying or renaming job configs, jobs
that were intended to only ever be run manually always get enabled on
every PR and nobody ever notices before merging the configuration
change.

Allow setting "always_run: false" in the config so that we at least have
the option to set it in the config and thus have it preserved when jobs
are renamed or duplicated.