test/e2e: remove source code lines from output
E2E tests are failing continually:

https://prow.ci.openshift.org/job-history/gs/origin-ci-test/pr-logs/directory/pull-ci-openshift-ci-tools-master-e2e

This is because the tested error output includes a source code line number,
causing it to fail whenever there is a change in `entrypoint`.  The output is
already changed by the test so that time differences are ignored, this change
also removes line numbers in a similar manner.

It uses a simple regular expression to turn messages such as

```
&#34;file&#34;:&#34;k8s.io/test-infra/prow/entrypoint/run.go:80&#34;
```

into

```
&#34;file&#34;:&#34;k8s.io/test-infra/prow/entrypoint/run.go&#34;
```

(where `&#34;` is just XML nonsense for the `"` character, the original message
is a JSON object).  There are other potential sources of instability in the
output, but this will at least get the back to its less-unstable state, we can
address those (or, more likely, the entire test) later.