Only include disruption NURPs if we have 100 rows.
We fail tests if you're over the P99 value, but it turns out the
majority of our platforms we were enforcing on do not have 100 results
per 3 weeks, so we weren't really getting P99s and thus could fail quite
often for some.

Update the query to include the count of JobRuns, and exclude if there's less
than 100.

Also added sorting so the file stays in a consistent order, which will
help us decipher diffs when they land in origin.

TRT-583