integration: regenerate `pj-rehearse` jobs
Otherwise these files quickly become obsolete and no longer test what is
currently being used in `openshift/release` (there are several examples in this
commit and in the `git` history in general).

The only minor problem is the job used for testing changes to the `ci-operator`
parameters can no longer be an automatically generated one, since that precludes
changing the parameters directly.