Enable setting cluster-claim in workflows
Up until now, cluster-claim stanza was only applicable on the job-config
level. It is sometimes desired to configure the cluster-claim parameters
in one place and reuse it across different job definitions.

This PR adds a propagation of cluster-claim from a workflow if it's
defined, with respect to it being overridden by job configuration if it
exists.