pkg/util: use a watch to wait for pods
The verification is implemented as a Go routine executed in parallel with the
`WaitForConditionOnObject` call.  While this could be done by using a `watch`
call directly, handling both event streams in a `for`/`select` expression, it
would require reimplementing much of the work done by the functions in
`k8s.io/client-go/tools/watch`.

In the end, it was simpler to start the verification as soon as an object is
available (after the initial synchronization done by the informer).  The last
object state seen by the watcher is kept and referenced by a pointer shared
with the pending check so that additional `get` calls are not necessary.
Atomic operations are used since there is only ever a single reader/writer
(weaker operations could be used in this case, but Go does not support those).