pkg/util: improve pod pending check
The goal of this function is to prevent a pod destined to be permanently pending
from delaying a test until the end of its configured timeout period (usually
several hours).

Any container can potentially be the cause of the problem, the current
distinction between infrastructure/test ones is not very useful.  The new
implementation checks all of them, at every point in the execution of the pod.