entrypoint-wrapper: create Git configuration
Newer versions of Git refuse to interact with a repository owned by another
user, independently of file permissions.  This is a problem in CI since source
code is usually cloned by UID 0 and read by unprivileged users.

With this change, `entrypoint-wrapper` creates a Git configuration file in the
home directory if it does not already exist (n.b.: this directory is already
provided as a temporary `EmptyDir` volume to `entrypoint-wrapper` in multi-stage
tests).  This file will disable the ownership verification using the following
entry:

    [safe]
        directory = *

Tests which want to have better control over their configuration can rewrite
this file or use one of the other configuration methods in Git:

- https://git-scm.com/docs/git-config.html#FILES
- https://git-scm.com/docs/git-config.html#SCOPES